\name{Hybrid EM-PAV Algorithm}
\alias{hybrid.em.pav}
\alias{expectation}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Hybrid EM-PAV Algorithm
}
\description{
The function \code{hybrid.em.pav} executes the hybrid expectation-maximization pool-adjacent-violators algorithm given in the paper "Misclassified Group Tested Current Status Data." The function \code{expectation} computes the "E" step in the EM-PAV hybrid algorithm. The iterations are done by the \code{hybrid.em.pav} function.
}
\usage{
hybrid.em.pav(data, initial.p, threshold=0.01, alpha=1, beta=1)
expectation(p, delta, alpha=1, beta=1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
This is a dataset as generated by the function \code{grouped_csdata_weibull.R}
}
  \item{initial.p}{
These are the initial values for the algorithm for each individual in the dataset. Can come from an external source, or should indicate which column of \code{data} contains the initial values.
}
  \item{threshold}{
Point at which algorithm should stop. Is the sum of the squared distances between last iteration and current iteration. Default is 0.01.
}
  \item{p}{
Probabilities to be updated by the \code{expectation} function. This is a vector the length of the group.
}
  \item{delta}{
  A scalar indicating the group test result.
  }
  \item{alpha}{
Sensitivity. Default is 1.
}
  \item{beta}{
Specificity. Default is 1.
}
}
\details{
This is an implementation of the EM algorithm for misclassified grouped current status data, using the pool-adjacent-violators algorithm (Ayers et al., 1955) as the maximization step. 
}
\value{
\item{results}{These should be treated as the right-continuous step function values (like the \code{x} value in the output of \code{gpava} from the \code{Isotone} package.) }

\item{num.iterations}{The total number of iterations until convergence.}

\item{diff}{The sum of the squared differences between the final iteration and the previous iteration. This is always less than the specified threshold.}
}
\references{

}
\author{
Lucia Petito
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
data <- gen.data.weibull.unif(100, 2, 4, 25, 25, 1, 1)
out <- hybrid.em.pav(data, data$initial.p, alpha=0.9, beta=0.9)
head(out$results)
out$num.iterations
out$diff
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{survival}
